pyproject.toml
[tool.poetry]
name = "contacts-api"
version = "0.1.0"
description = "REST API for contact management"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.22"
psycopg2-binary = "^2.9.9"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
python-dotenv = "^1.0.0"
alembic = "^1.12.0"
email-validator = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
fastapi-limiter = "^0.1.5"
redis = "^5.0.1"
fastapi-cors = "^0.0.6"

[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contacts-api"
version = "0.1.0"
description = "REST API for contacts management"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "sqlalchemy>=1.4.23",
    "pydantic>=1.8.2",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.5",
    "alembic>=1.7.1",
    "psycopg2-binary>=2.9.1",
    "redis>=4.0.0",
    "fastapi-limiter>=0.1.5",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.18.0",
    "httpx>=0.24.0",
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "sphinx-autodoc-typehints>=1.23.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=app --cov-report=term-missing"

[tool.coverage.run]
source = ["app"]
omit = ["tests/*", "docs/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
] 